Create make_ffnet
The function net=make_ffnet(Nlayers, Nneurons, hasBiasNeuron)
It creates and initializes a feed-forward neural network with Nlayers layers. The number of (non-bias) neurons in each layer is given by the vector Nneurons and whether a layer has a bias neuron or not is determined by the logical vector hasBiasNeuron. These two vectors must be of size Nlayers x 1. The output of the function is a structure with the fields.
Nlayers - the number of layers
Nneurons - the vector with the number of neurons in each layer
hasBiasNeuron - the vector determining whether a layer has a bias neuron O a cell array such that net.O{l}(j) is the value of the neuron output for neuron j in layer l. Neuron number Nneurons(l)+1 is the bias neuron (if there is any).
I - a cell array such that net.I{l}(j) is the weighted sum of inputs to neuron j in layer l
factiv - a cell array such that net.factiv{l} is a function handle to the activation function used in layer l. Use tanh for the output layer and ReLU for all other layers
w - It is the weight between neuron j of layer l and neuron k of layer l + 1. The weights are initialized to normally distributed random numbers with variance 2=(M + N), where M is the number of neurons in layer l, and N is the number of neurons in layer l + 1
E - function handle to a function E(y; ^y), that computes how well the objective was achieved, if the output vector is y and the true result ^y. 
dE - function handle to a function that computes the partial derivatives of E with respect to y1....yN